+++ Sources/Extensions/SwiftStdlib/StringExtensions.swift
@@ -105,7 +105,7 @@ public extension String {
 		return String(first)
 	}
 
-	#if canImport(Foundation)
+    #if canImport(Foundation)
 	/// SwifterSwift: Check if string contains one or more letters.
 	///
 	///		"123abc".hasLetters -> true
@@ -636,6 +636,16 @@ public extension String {
 
 		self = first + rest
 	}
+
+    /// SwifterSwift: First character of string uppercased(if applicable) while keeping the original string.
+    ///
+    ///        "hello world".firstCharacterUppercased() -> "Hello world"
+    ///        "".firstCharacterUppercased() -> ""
+    ///
+    public mutating func firstCharacterUppercased() {
+        guard let first = first else { return }
+        self = String(first).uppercased() + dropFirst()
+    }
 
 	/// SwifterSwift: Check if string contains only unique characters.
 	///